clear some compiling warnings

+++ Sources/Elastos/LibCore/inc/elastos/core/NativeThread.h
@@ -29,7 +29,7 @@ typedef struct NativeObject
      * A word containing either a "thin" lock or a "fat" monitor.  See
      * the comments in Sync.c for a description of its layout.
      */
-    UInt32      mLock;
+    size_t      mLock;
     Int64       mObjectObj;
 } NativeObject;
